+++ src/main/java/devcourse/springbootbasic/commandline/function/FunctionHandler.java
@@ -7,10 +7,13 @@ import devcourse.springbootbasic.commandline.domain.VoucherTypeSelector;
 import devcourse.springbootbasic.controller.CustomerController;
 import devcourse.springbootbasic.controller.VoucherController;
 import devcourse.springbootbasic.domain.voucher.VoucherType;
-import devcourse.springbootbasic.dto.CustomerFindResponse;
-import devcourse.springbootbasic.dto.VoucherCreateRequest;
-import devcourse.springbootbasic.dto.VoucherCreateResponse;
-import devcourse.springbootbasic.dto.VoucherFindResponse;
+import devcourse.springbootbasic.dto.customer.CustomerCreateRequest;
+import devcourse.springbootbasic.dto.customer.CustomerCreateResponse;
+import devcourse.springbootbasic.dto.customer.CustomerFindResponse;
+import devcourse.springbootbasic.dto.customer.CustomerUpdateResponse;
+import devcourse.springbootbasic.dto.voucher.VoucherCreateRequest;
+import devcourse.springbootbasic.dto.voucher.VoucherCreateResponse;
+import devcourse.springbootbasic.dto.voucher.VoucherFindResponse;
 import devcourse.springbootbasic.exception.InputErrorMessage;
 import devcourse.springbootbasic.exception.InputException;
 import lombok.RequiredArgsConstructor;
