+++ src/main/java/devcourse/springbootbasic/controller/CustomerController.java
@@ -1,6 +1,9 @@
 package devcourse.springbootbasic.controller;
 
-import devcourse.springbootbasic.dto.CustomerFindResponse;
+import devcourse.springbootbasic.dto.customer.CustomerCreateRequest;
+import devcourse.springbootbasic.dto.customer.CustomerCreateResponse;
+import devcourse.springbootbasic.dto.customer.CustomerFindResponse;
+import devcourse.springbootbasic.dto.customer.CustomerUpdateResponse;
 import devcourse.springbootbasic.service.CustomerService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
